# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "API автоматизации рабочих процессов",
        "contact": {},
        "version": "1"
    },
    "host": "localhost:8080/",
    "paths": {
        "/auth/token": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Генерирует токен авторизации",
                "parameters": [
                    {
                        "description": "Данные для генерации токена",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth_handler.TokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Response-auth_handler_TokenResponse"
                        }
                    },
                    "401": {
                        "description": "Не известный пользователь",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "422": {
                        "description": "Ошибка в запросе",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            }
        },
...
        "/management/reviews": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "code-reviews"
                ],
                "summary": "Получить список всех код-ревью",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Response-array_code_review_handler_CodeReviewResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "code-reviews"
                ],
                "summary": "Создать новое код-ревью",
                "parameters": [
                    {
                        "description": "Данные код-ревью",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/code_review_handler.CreateCodeReviewRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Response-code_review_handler_CodeReviewResponse"
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            }
        },
        "/management/reviews/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "code-reviews"
                ],
                "summary": "Получить информацию о код-ревью",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID код-ревью",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Response-code_review_handler_CodeReviewResponse"
                        }
                    },
                    "404": {
                        "description": "Код-ревью не найдено",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации ID",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "code-reviews"
                ],
                "summary": "Обновить информацию о код-ревью",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID код-ревью",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новые данные код-ревью",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/code_review_handler.UpdateCodeReviewRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Response-code_review_handler_CodeReviewResponse"
                        }
                    },
                    "404": {
                        "description": "Код-ревью не найдено",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "code-reviews"
                ],
                "summary": "Удалить код-ревью",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID код-ревью",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Код-ревью не найдено",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации ID",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.Error": {
            "description": "Модель ошибки на запрос",
            "type": "object",
            "properties": {
                "message": {
                    "description": "Текст ошибки",
                    "type": "string"
                }
            }
        },
        "api.Response-any": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Данные"
                },
                "error": {
                    "description": "Ошибка",
                    "allOf": [
                        {
                            "$ref": "#/definitions/api.Error"
                        }
                    ]
                }
            }
        },
        "api.Response-array_code_review_handler_CodeReviewResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Данные",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/code_review_handler.CodeReviewResponse"
                    }
                },
                "error": {
                    "description": "Ошибка",
                    "allOf": [
                        {
                            "$ref": "#/definitions/api.Error"
                        }
                    ]
                }
            }
        },
...
        "api.Response-auth_handler_TokenResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Данные",
                    "allOf": [
                        {
                            "$ref": "#/definitions/auth_handler.TokenResponse"
                        }
                    ]
                },
                "error": {
                    "description": "Ошибка",
                    "allOf": [
                        {
                            "$ref": "#/definitions/api.Error"
                        }
                    ]
                }
            }
        },
        "api.Response-code_review_handler_CodeReviewResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Данные",
                    "allOf": [
                        {
                            "$ref": "#/definitions/code_review_handler.CodeReviewResponse"
                        }
                    ]
                },
                "error": {
                    "description": "Ошибка",
                    "allOf": [
                        {
                            "$ref": "#/definitions/api.Error"
                        }
                    ]
                }
            }
        },
...
        "auth_handler.TokenRequest": {
            "description": "запрос на генерацию токена",
            "type": "object",
            "properties": {
                "password": {
                    "description": "@Description Пароль",
                    "type": "string"
                },
                "username": {
                    "description": "@Description Имя пользователя",
                    "type": "string"
                }
            }
        },
        "auth_handler.TokenResponse": {
            "description": "Модель со сгенерированным токеном",
            "type": "object",
            "properties": {
                "token": {
                    "description": "@Description токен",
                    "type": "string"
                }
            }
        },
        "code_review_handler.CodeReviewResponse": {
            "description": "Информация о процессе код-ревью",
            "type": "object",
            "properties": {
                "id": {
                    "description": "@Description\tУникальный ID\n\t@Example:\t\t1",
                    "type": "integer"
                },
                "pull_request_number": {
                    "description": "@Description\tНомер Pull Request\n\t@Example:\t\t\"42\"",
                    "type": "string"
                },
                "status": {
                    "description": "@Description\tСтатус код-ревью\n\t@Example:\t\t\"approved\"\n\t@Enum\t\t\tpending,approved,rejected",
                    "type": "string"
                },
                "thread_ts": {
                    "description": "@Description\tИдентификатор треда в Slack\n\t@Example:\t\t\"1234567890.123456\"",
                    "type": "string"
                }
            }
        },
        "code_review_handler.CreateCodeReviewRequest": {
            "description": "Запрос на создание код-ревью",
            "type": "object",
            "required": [
                "pull_request_number",
                "status",
                "thread_ts"
            ],
            "properties": {
                "pull_request_number": {
                    "description": "@Description\tНомер Pull Request\n\t@Example:\t\t\"42\"\n\t@Required\t\ttrue",
                    "type": "string"
                },
                "status": {
                    "description": "@Description\tСтатус код-ревью\n\t@Example:\t\t\"pending\"\n\t@Enum\t\t\tpending,approved,rejected\n\t@Required\t\ttrue",
                    "type": "string",
                    "enum": [
                        "pending",
                        "approved",
                        "rejected"
                    ]
                },
                "thread_ts": {
                    "description": "@Description\tИдентификатор треда в Slack\n\t@Example:\t\t\"1234567890.123456\"\n\t@Required\t\ttrue",
                    "type": "string"
                }
            }
        },
        "code_review_handler.UpdateCodeReviewRequest": {
            "description": "Запрос на обновление код-ревью",
            "type": "object",
            "properties": {
                "pull_request_number": {
                    "description": "@Description\tНомер Pull Request\n\t@Example:\t\t\"42\"",
                    "type": "string"
                },
                "status": {
                    "description": "@Description\tСтатус код-ревью\n\t@Example:\t\t\"approved\"\n\t@Enum\t\t\tpending,approved,rejected",
                    "type": "string",
                    "enum": [
                        "pending",
                        "approved",
                        "rejected"
                    ]
                },
                "thread_ts": {
                    "description": "@Description\tИдентификатор треда в Slack\n\t@Example:\t\t\"1234567890.123456\"",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}